refactor: shutdown function in pkg for both servers
