+++ pkg/shutdown/shutdown.go
@@ -1,0 +1,41 @@
+package pkg
+
+import (
+	"context"
+	"log"
+	"net/http"
+	"os"
+	"os/signal"
+	"syscall"
+	"time"
+)
+
+func ShutDown(server *http.Server) {
+
+	// Create a channel to receive signals
+	signalCh := make(chan os.Signal, 1)
+	signal.Notify(signalCh, syscall.SIGINT, syscall.SIGTERM)
+
+	// Start the server in a separate goroutine
+	go func() {
+		log.Printf("Server listening on %s\n", server.Addr)
+		if err := server.ListenAndServe(); err != nil && err != http.ErrServerClosed {
+			log.Fatalf("listen: %s\n", err)
+		}
+	}()
+
+	// Wait for a signal to shutdown the server
+	sig := <-signalCh
+	log.Printf("Received signal: %v\n", sig)
+
+	// Create a context with a timeout
+	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
+	defer cancel()
+
+	// Shutdown the server gracefully
+	if err := server.Shutdown(ctx); err != nil {
+		log.Fatalf("Server shutdown failed: %v\n", err)
+	}
+
+	log.Println("Server shutdown gracefully")
+}
